Platformer minigame update: * added distance tracking and obstacle collisions * improved the map handling process
We'll probably need to draw a frame/border around the game area to
hide the tiles that appear/disappear while scrolling the map. This, 
or figure a way to make the minigame ui stay on top of the minigame
area. Applies to all games/puzzles that use the minigame framework.