Battle update.
Added a simple system to allow the scripter to decide whether a 
battle should be fought to the end or if the player would be 
allowed to quit the battle by using the exit button. Happened 
to rewrite most of the battle loop while doing this, adding game 
states and probably even making the code a bit more readable in 
the process.