Search update.
Added a few lines of comments, and while doing that, noticed that
entering a room from the map view removes the currently active
background (story/main view, whatever you call it). This is because
the room images are defined as backgrounds, too (bg riroom etc.), and
thus replace the real background when used. This can be fixed by
removing the "bg" part from the room names, so that Renpy can treat
them as normal images rather than backgrounds.