Minigame Squats update.
More or less an overhaul to make the minigame be more like the one described in 
the GDD. There are still some differences with this and the one in the document 
(e.g., this version simply checks whether or not you managed to hit the green or 
the orange bar, ignoring any full reps you manage to do. Some bonus for that 
could be nice, though.), but it gets the job done.