Minigame Cells update.
Changed the basic game mechanics a bit to make the minigame easier to play and tweak (and 
more fun just to watch). Now the game starts with only a few healthy cells in the dish. Once 
those cells have duplicated themselves for some time, mutations kick in and start creating 
hostile cells that attack their neighbors and spread the infection to any adjacent healthy 
cells. One cell can mutate a total of three times until it is of the hardest type possible.

Another big but less visible change is the improvement in the minigame's overall 
performance. Even though it still uses a lot of resources, it isn't quite that taxing on 
older machines either (according to the significantly dropped core load and fan noise 
levels). While still far from optimal, it's a leap in the right direction.